+++ src/components/Grades/TabViewButtons.js
@@ -1,15 +1,21 @@
 import React from 'react'
+import styled from 'styled-components';
 
 import AllPeriodsButton from '../../parts/Buttons/AllPeriodsButton'
 import PeriodButton from '../../parts/Buttons/PeriodButton'
 
+const Container = styled.div`
+    display: flex;
+    flex-direction: column;
+
+    > * {
+        margin: 7px 0 0 0;
+    }
+`
+
 export default function TabViewButtons({setPeriod, periods, selectedPeriod}) {
     return (
-        <div style={{
-            display: 'flex',
-            overflow: 'auto',
-            height: '48px'
-        }}>
+        <Container>
             <AllPeriodsButton setPeriod={setPeriod} selected={!selectedPeriod}/>
             {
                 periods.map((period, index) => {
@@ -22,6 +28,6 @@ export default function TabViewButtons({setPeriod, periods, selectedPeriod}) {
                     />
                 })
             }
-        </div>
+        </Container>
     )
 }
