+++ src/parts/Cards/SubjectCard.js
@@ -6,23 +6,22 @@ import getGradeColor from '../../common/getGradeColor';
 const Card = styled.div`
     height: 100px;
     display: grid;
-    width: 295px;
     grid-template-columns: 100px 1fr;
     background-color: #1E1F2F;
     border-radius: 10px;
-    margin: 5px;
+    margin: 5px 0;
     cursor: pointer;
 `
 
 export default function SubjectCard({subject, onClick}) {
     return (
         <Card onClick={onClick}>
-            <div style={{position: 'relative'}}>
+            <div style={{position: 'relative', display: 'grid', placeItems: 'center'}}>
                 <RoundReadOnlySlider
                     value={subject.averageGrade}
                     progressColor={getGradeColor(subject.averageGrade)}
-                    size={100}
-                    progressWidth={8}
+                    size={90}
+                    progressWidth={10}
                 />
                 <p style={{
                     position: 'absolute',
